feat: location setting in SettingLocationActivity
- we can choose my location in two options
- also set locationNear info
- when return back to MainActivity(HomeFragment), item datas will be changed according to your selected info

+++ app/src/main/java/com/dongldh/carrotmarket/database/Datas.kt
@@ -10,7 +10,7 @@ data class DataUser(var phone: String = "test",
                     var userName: String = "로그인하세요",
                     var location: ArrayList<String?> = arrayListOf(""),
                     var profileImage: String? = null,
-                    var locationNear: ArrayList<Long?> = arrayListOf(1L))
+                    var locationNear: ArrayList<Int?> = arrayListOf(1))
 
 // 상품 등록 시 필요한 데이터 저장
 // type : 1 -> 중고물품 거래,  2 -> 지역 홍보
